docs(dreamlite): address final review nits from #13815
- Replace broken cat.png URL in editing examples (both base and mobile)
  with the standard `huggingface/documentation-images` source used
  elsewhere in the diffusers docs.
- Promote the recommended guidance_scale=3.5 / image_guidance_scale=1.5
  to the default values of DreamLitePipeline.__call__, and drop the
  now-redundant explicit args from the docs examples.
- Switch the EXAMPLE_DOC_STRING examples in both pipelines from
  torch.float16 to torch.bfloat16 for consistency with the rest of the
  docs.

+++ src/diffusers/pipelines/dreamlite/pipeline_dreamlite_mobile.py
@@ -139,7 +139,7 @@ EXAMPLE_DOC_STRING = """
         >>> from diffusers import DreamLiteMobilePipeline
 
         >>> pipe = DreamLiteMobilePipeline.from_pretrained(
-        ...     "carlofkl/DreamLite-mobile", revision="diffusers", torch_dtype=torch.float16
+        ...     "carlofkl/DreamLite-mobile", revision="diffusers", torch_dtype=torch.bfloat16
         ... )
         >>> pipe.to("cuda")
